+++ MobileProject2021/Assets/Scripts/PlayerMechanics/Upgrades/OnUpgradeEventArgs.cs
@@ -5,10 +5,10 @@ using UnityEngine;
 
 public class OnUpgradeEventArgs : EventArgs
 {
-    public int currencyAmount;
-    public int currencyBase;
-    public int currencyPremium;
-    public int currencyElemental;
+    public float currencyAmount;
+    public float currencyBase;
+    public float currencyPremium;
+    public float currencyElemental;
 
     public float upgradeEffect;
     public float upgradeLevel;
